♻️ refactoring : 인터뷰 타이핑으로 인해 발생되는 불필요한 리렌더링 방지 및 수정
